demo 1&2 improvements
