+++ src/test/java/graphicsEngineDemo/d1_simplestDemo/ReadMe.md
@@ -4,6 +4,8 @@
 
 This is a demo demonstrating the simplest possible use of the Graphics Engine.
 
+Starts a blank SimpleWindow.
+
 Everything contained in just one class.
 
 ## Instructions
