+++ src/test/java/graphicsEngineDemo/d2_overlayDemo/Overlay.java
@@ -1,11 +1,13 @@
 package graphicsEngineDemo.d2_overlayDemo;
 
-import graphicsEngine.presets.SimpleOverlay;
-
 import java.awt.event.ActionListener;
 import javax.swing.BoxLayout;
 
 import org.jetbrains.annotations.Nullable;
+
+import graphicsEngine.presets.SimpleOverlay;
+
+import graphicsEngineDemo.d2_overlayDemo.buttons.Button4;
 
 /**
  * TODO: finish this javadoc
@@ -15,6 +17,6 @@ class Overlay extends SimpleOverlay {
     protected Overlay(@Nullable ActionListener actionListener) {
         super(OverlayColors.RED);
         setLayout(new BoxLayout(this, BoxLayout.X_AXIS));
-        add(new ButtonListener.Button4(actionListener));
+        add(new Button4(actionListener));
     }
 }
