+++ src/test/java/graphicsEngineDemo/d2_overlayDemo/ReadMe.md
@@ -4,6 +4,10 @@
 
 This is a demo for demonstrating the overlay functionality.
 
+Launches a SinglePageWindow with an overlay.
+
+Contains buttons for toggling overlay visibility and color.
+
 ## Instructions
 
 Run `Main.main(String[] args)` to launch this demo.
